Rename `model.proto` to `tasks.proto`
