+++ client-angular/test/app/given/task-creation-process.ts
@@ -21,7 +21,7 @@
 import {chore} from 'test/given/tasks';
 
 import {TaskCreationId} from 'proto/todolist/identifiers_pb';
-import {TaskCreation} from 'proto/todolist/model_pb';
+import {TaskCreation} from 'proto/todolist/tasks_pb';
 import {LabelIdsList} from 'proto/todolist/values_pb';
 import {LabelView, TaskView} from 'proto/todolist/projections_pb';
