+++ client-angular/test/app/task-creation-wizard/service/task-creation-wizard.service.spec.ts
@@ -32,7 +32,7 @@ import {tomorrow, yesterday} from 'test/given/dates';
 
 import {Timestamp} from 'google-protobuf/google/protobuf/timestamp_pb';
 import {TaskPriority} from 'proto/todolist/attributes_pb';
-import {TaskCreation} from 'proto/todolist/model_pb';
+import {TaskCreation} from 'proto/todolist/tasks_pb';
 import {TaskDescription} from 'proto/todolist/values_pb';
 import {TaskView} from 'proto/todolist/projections_pb';
 import {LayoutModule} from 'app/layout/layout.module';
